+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 🔥🔥🔥 Version 0.4.0 🔥🔥🔥
+### ✨ Add: 
+- Add new character in HSR v3.2: **Castorice & Anaxa**.
+- Add **Setting** and **Volume** button in HSR Ban/Pick UI like GI.
+- Add ban/pick turn notice on team name title.
+
+### 🔧 Fix:
+- Erase some unuse line.
+- Make time.js is a global tool for this project. 
+
 ## Version 0.3.0
 ### Add:
 - Add new character in GI v5.5: Varesa & Iansan.
@@ -13,7 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add ban/pick turn notice on team name title.
 
 ### Fix:
-- Change folder name from HUI to HUD.
+- Change folder name in HSR and GI from HUI to HUD.
 
 ## Version 0.2.2
 ### Fix: